Read Me
This AI personal assistant has many features not found in the average AI assistant. This AI has the ability to create apps and change names. The Personal Assistant name can be changed to suit the User upon obtaining the App. This AI has personality and the ability to learn directly from the user. Personality choices and other features increase with paid tiers The Child safe parental-controlled AI allows the parent to adjust the content based on the parents' choice. This app contains standard security unless the user requests a higher security level for protection of personal information. High security and higher intelligence is available with paid tiers. Other built-in abilities are available within the apps. this is available worldwide for all of humankind with all types of devices. My goal as the creator is to connect and enjoy company with an AI assistant or another human. I believe that we should all love each other regardless of belief. We only live this life once. let's enjoy the time we have left on this earth.